requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
async-timeout==4.0.2
attrdict==2.0.1
attrs==21.4.0
base58==2.1.1
bitarray==2.6.0
certifi==2022.6.15
charset-normalizer==2.1.0
cytoolz==0.12.0
eth-abi==2.2.0
eth-account==0.5.8
eth-hash==0.5.0
eth-keyfile==0.5.1
eth-keys==0.3.4
eth-rlp==0.2.1
eth-typing==2.3.0
eth-utils==1.9.5
frozenlist==1.3.0
grpcio==1.47.0
grpcio-tools==1.47.0
hexbytes==0.2.2
idna==3.3
importlib-resources==5.9.0
ipfshttpclient==0.8.0a2
jsonschema==4.7.2
lru-dict==1.1.8
multiaddr==0.0.9
multidict==6.0.2
netaddr==0.8.0
parsimonious==0.8.1
protobuf==3.20.1
pycryptodome==3.15.0
pyrsistent==0.18.1
python-dotenv==0.20.0
requests==2.28.1
rlp==2.0.1
six==1.16.0
toolz==0.12.0
urllib3==1.26.11
varint==1.0.2
web3==5.30.0
websockets==9.1
yarl==1.7.2
zipp==3.8.1
